Rename api resources
