+++ web/app/api_1_0/__init__.py
@@ -2,4 +2,4 @@ from flask import Blueprint
 
 api = Blueprint('api', __name__)
 
-from . import api_exception, api_map, game_result, game_server, player_info
+from . import api_exception, api_map, game_resource, goserver_resource, player_resource
