+++ web/app/api_1_0/goserver_resource.py
