+++ web/app/api_1_0/player_resource.py
